requirements.txt
Flask
pyrebase4
python-dotenv